Moves `FeedImagesMapper` to its own file.
