+++ FeedAPIChallenge/Feed API/FeedImagesMapper.swift
@@ -1,0 +1,39 @@
+//
+//  Copyright © 2020 Essential Developer Ltd. All rights reserved.
+//
+
+import Foundation
+
+internal final class FeedImagesMapper {
+	private static let OK_200 = 200
+	
+	private struct Root: Decodable {
+		private let items: [Image]
+		var feed: [FeedImage] { items.map{ $0.feedImage } }
+	}
+	
+	private struct Image: Decodable {
+		private let id: UUID
+		private let description: String?
+		private let location: String?
+		private let url: URL
+		
+		var feedImage: FeedImage {
+			FeedImage(id: id, description: description, location: location, url: url)
+		}
+		
+		private enum CodingKeys: String, CodingKey {
+			case id = "image_id"
+			case description = "image_desc"
+			case location = "image_loc"
+			case url = "image_url"
+		}
+	}
+	
+	static func map(data: Data, response: HTTPURLResponse) -> FeedLoader.Result {
+		guard response.statusCode == OK_200, let root = try? JSONDecoder().decode(Root.self, from: data) else {
+			return .failure(RemoteFeedLoader.Error.invalidData)
+		}
+		return .success(root.feed)
+	}
+}
